requirements.txt
opencv_python==3.4.1.15
scipy==1.2.3
matplotlib==2.1.1
scikit_image==0.14.5
numpy==1.13.3
open3d==0.12.0
Pillow==8.1.0
scikit_learn==0.24.0
skimage==0.0
